+++ CHANGELOG.md
@@ -43,6 +43,8 @@
   end
   ```
 
+- Made FactoryProf measure and report on timing ([@danielwaterworth][])
+
 ## 0.8.0 (2019-04-12) 🚀
 
 - **Ruby 2.4+ is requiered** ([@palkan][])